Switch back to ID=raspbian in /etc/os-release
In order to make HypriotOS for RPi compatible with Docker Machine, we had to modify the ID from
`raspbian` to `debian` in the past. This was just a hack which should be obsolete as soon as the new raspbian provisioner will be merged into Docker Machine with this PR https://github.com/docker/machine/pull/3605.

Signed-off-by: Dieter Reuter <dieter.reuter@me.com>